+++ backend/schedule/tests/test_admin.py
@@ -1,7 +1,6 @@
 from unittest.mock import call
 
 import pytest
-import time_machine
 from django.utils import timezone
 
 from conferences.models import SpeakerVoucher
@@ -10,122 +9,10 @@ from schedule.admin import (
     send_schedule_invitation_reminder_to_waiting,
     send_schedule_invitation_to_all,
     send_schedule_invitation_to_uninvited,
-    send_voucher_via_email,
 )
 from schedule.models import ScheduleItem
 
 pytestmark = pytest.mark.django_db
-
-
-@time_machine.travel("2020-10-10 10:00:00", tick=False)
-def test_send_voucher_via_email(
-    rf,
-    schedule_item_factory,
-    conference_factory,
-    submission_factory,
-    speaker_voucher_factory,
-    mocker,
-):
-    mocker.patch("schedule.admin.messages")
-    mock_send_email = mocker.patch("schedule.admin.send_speaker_voucher_email")
-
-    conference = conference_factory(pretix_speaker_voucher_quota_id=123)
-    schedule_item_factory(
-        type=ScheduleItem.TYPES.submission,
-        conference=conference,
-        submission=submission_factory(conference=conference, speaker_id=500),
-    )
-    schedule_item_factory(
-        type=ScheduleItem.TYPES.submission,
-        conference=conference,
-        submission=submission_factory(conference=conference, speaker_id=600),
-    )
-
-    speaker_voucher_1 = speaker_voucher_factory(
-        conference=conference,
-        user_id=500,
-        pretix_voucher_id=1,
-    )
-    speaker_voucher_2 = speaker_voucher_factory(
-        conference=conference,
-        user_id=600,
-        pretix_voucher_id=2,
-    )
-
-    send_voucher_via_email(
-        None, rf.get("/"), queryset=SpeakerVoucher.objects.filter(conference=conference)
-    )
-
-    mock_send_email.assert_has_calls(
-        [
-            call(speaker_voucher_1),
-            call(speaker_voucher_2),
-        ]
-    )
-
-    speaker_voucher_1.refresh_from_db()
-    speaker_voucher_2.refresh_from_db()
-
-    assert speaker_voucher_1.voucher_email_sent_at == timezone.now()
-    assert speaker_voucher_2.voucher_email_sent_at == timezone.now()
-
-
-@time_machine.travel("2020-10-10 10:00:00", tick=False)
-def test_send_voucher_via_email_requires_filtering_by_conference(
-    rf,
-    schedule_item_factory,
-    conference_factory,
-    submission_factory,
-    speaker_voucher_factory,
-    mocker,
-):
-    mock_messages = mocker.patch("schedule.admin.messages")
-    mock_send_email = mocker.patch("schedule.admin.send_speaker_voucher_email")
-
-    conference = conference_factory(pretix_speaker_voucher_quota_id=123)
-    conference_2 = conference_factory(pretix_speaker_voucher_quota_id=123)
-
-    schedule_item_factory(
-        type=ScheduleItem.TYPES.submission,
-        conference=conference,
-        submission=submission_factory(conference=conference, speaker_id=500),
-    )
-    schedule_item_factory(
-        type=ScheduleItem.TYPES.submission,
-        conference=conference_2,
-        submission=submission_factory(conference=conference_2, speaker_id=600),
-    )
-
-    speaker_voucher_1 = speaker_voucher_factory(
-        conference=conference,
-        user_id=500,
-        pretix_voucher_id=1,
-    )
-    speaker_voucher_2 = speaker_voucher_factory(
-        conference=conference_2,
-        user_id=600,
-        pretix_voucher_id=2,
-    )
-
-    request = rf.get("/")
-    send_voucher_via_email(
-        None,
-        request,
-        queryset=SpeakerVoucher.objects.filter(
-            conference__in=[conference, conference_2]
-        ),
-    )
-
-    mock_messages.error.assert_called_once_with(
-        request, "Please select only one conference"
-    )
-    mock_send_email.assert_not_called()
-
-    speaker_voucher_1.refresh_from_db()
-    speaker_voucher_2.refresh_from_db()
-
-    assert speaker_voucher_1.voucher_email_sent_at is None
-    assert speaker_voucher_2.voucher_email_sent_at is None
 
 
 def test_generate_voucher_codes(
